refactor: extract theme toggling logic to a separate Provider
