+++ src/context/index.types.ts
@@ -1,5 +1,4 @@
-import { UserLocationType, SnackbarStateType, UserType } from '@/types';
-import { Theme } from '@mui/material';
+import { SnackbarStateType, UserLocationType, UserType } from '@/types';
 import { Session } from '@supabase/supabase-js';
 import { Dispatch, SetStateAction } from 'react';
 
@@ -10,10 +9,6 @@ export interface ContextProps {
   setRegisteringUser: Dispatch<SetStateAction<UserType | null>>;
   userLocation: UserLocationType;
   setUserLocation: Dispatch<SetStateAction<UserLocationType>>;
-  isDarkModeOn: boolean;
-  setIsDarkModeOn: Dispatch<SetStateAction<boolean>>;
-  toggleTheme: () => void;
-  theme: Theme;
   normalizedPath: string;
   avatarPath: string | null;
   uploadAvatar: (file: File, session: Session) => void;
